scripts/seo-agent/requirements.txt
pyyaml>=6.0
google-analytics-data>=0.19.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-genai>=1.0.0
requests>=2.28.0
